LayoutTests/fast/css/parsing-stroke-width-expected.txt

This tests checks that all of the input values for stroke-width parse correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testComputedStyleInherited("stroke-width: 4px;") is "4px"
FAIL testComputedStyleInherited("stroke-width: 0.01em;") should be 0.01em. Was 0.16px.
PASS testComputedStyleInherited("stroke-width: 10%;") is "10%"
PASS testComputedStyle(";") is "1px"
PASS test("stroke-width: 4px;") is "4px"
PASS test("stroke-width: 0.01em;") is "0.01em"
PASS test("stroke-width: 10%;") is "10%"
PASS test("stroke-width: 4;") is "4"
PASS test("stroke-width: em;") is ""
PASS test("stroke-width: %;") is ""
PASS successfullyParsed is true

TEST COMPLETE
